+++ consensus/cryptarchia-engine/src/block.rs
@@ -1,4 +1,4 @@
-use crate::{crypto::Blake2b, time::Slot};
+use crate::{crypto::Blake2b, leader_proof::LeaderProof, time::Slot};
 use blake2::Digest;
 
 #[derive(Clone, Debug, Eq, PartialEq, Copy, Hash)]
@@ -9,68 +9,6 @@ pub struct ContentId([u8; 32]);
 
 #[derive(Clone, Debug, Eq, PartialEq, Copy)]
 pub struct Nonce([u8; 32]);
-
-#[derive(Clone, Debug, Eq, PartialEq, Copy, Hash)]
-pub struct LeaderProof {
-    commitment: Commitment,
-    nullifier: Nullifier,
-    slot: Slot,
-    evolved_commitment: Commitment,
-}
-
-impl LeaderProof {
-    pub fn commitment(&self) -> &Commitment {
-        &self.commitment
-    }
-
-    pub fn nullifier(&self) -> &Nullifier {
-        &self.nullifier
-    }
-
-    pub fn slot(&self) -> Slot {
-        self.slot
-    }
-
-    pub fn evolved_commitment(&self) -> &Commitment {
-        &self.evolved_commitment
-    }
-
-    #[cfg(test)]
-    pub fn dummy(slot: Slot) -> Self {
-        Self {
-            commitment: Commitment([0; 32]),
-            nullifier: Nullifier([0; 32]),
-            slot,
-            evolved_commitment: Commitment([0; 32]),
-        }
-    }
-
-    pub fn new(
-        commitment: Commitment,
-        nullifier: Nullifier,
-        slot: Slot,
-        evolved_commitment: Commitment,
-    ) -> Self {
-        Self {
-            commitment,
-            nullifier,
-            slot,
-            evolved_commitment,
-        }
-    }
-}
-
-#[derive(Clone, Debug, Eq, PartialEq, Copy, Hash)]
-pub struct Commitment([u8; 32]);
-
-#[derive(Clone, Debug, Eq, PartialEq, Copy, Hash)]
-pub struct Nullifier([u8; 32]);
-
-impl Nullifier {
-    pub fn as_bytes(&self) -> &[u8] {
-        &self.0
-    }
-}
 
 #[derive(Clone, Debug, Eq, PartialEq, Hash)]
 pub struct Header {
@@ -96,9 +34,9 @@ impl Header {
         h.update(self.slot.to_be_bytes());
         h.update(self.parent.0);
 
-        h.update(self.leader_proof.commitment.0);
-        h.update(self.leader_proof.nullifier.0);
-        h.update(self.leader_proof.evolved_commitment.0);
+        h.update(self.leader_proof.commitment());
+        h.update(self.leader_proof.nullifier());
+        h.update(self.leader_proof.evolved_commitment());
 
         for proof in &self.orphaned_leader_proofs {
             proof.update_hasher(h)
@@ -200,29 +138,5 @@ impl From<[u8; 32]> for ContentId {
 impl From<ContentId> for [u8; 32] {
     fn from(id: ContentId) -> Self {
         id.0
-    }
-}
-
-impl From<[u8; 32]> for Commitment {
-    fn from(commitment: [u8; 32]) -> Self {
-        Self(commitment)
-    }
-}
-
-impl From<Commitment> for [u8; 32] {
-    fn from(commitment: Commitment) -> Self {
-        commitment.0
-    }
-}
-
-impl From<[u8; 32]> for Nullifier {
-    fn from(nullifier: [u8; 32]) -> Self {
-        Self(nullifier)
-    }
-}
-
-impl From<Nullifier> for [u8; 32] {
-    fn from(nullifier: Nullifier) -> Self {
-        nullifier.0
     }
 }
